// frontend/src/app/api/generated/model/customerCreate.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CustomerCreate {
  city?: string | null;
  confirm_possible_duplicate?: boolean;
  country?: string | null;
  date_of_birth?: string | null;
  email?: string | null;
  first_name: string;
  house_number?: string | null;
  last_name: string;
  phone?: string | null;
  postal_code?: string | null;
  street?: string | null;
}